+++ src/test/java/com/github/yungyu16/spring/stub/example/service/InvocationDispatcherImpl1.java
@@ -2,7 +2,6 @@ package com.github.yungyu16.spring.stub.example.service;
 
 import com.github.yungyu16.spring.stub.example.annotation.TestClient;
 import com.github.yungyu16.spring.stub.proxy.AbstractInvocationDispatcher;
-import com.github.yungyu16.spring.stub.proxy.StubProxyContext;
 import org.springframework.stereotype.Component;
 
 import java.lang.reflect.Method;
